Refactoring to decouple rawlog management from gui
* Created CObservationTree and CObservationTreeItem classes to manage
the observations loaded from the rawlog, independent of gui, and their
synchronization.
* Subclassed CObservationTree as CObservationTreeGui to display the tree
in gui.
* Running plane segmentation now directly on the synced observation
tree. But display causes a crash.
* Fixed missing image for input2 issue, and the issue of incorrectly transforming
point clouds before display.
* Yet to add QProgressDialog.

+++ gui/interfaces/CRtObserver.h
@@ -9,5 +9,5 @@
 class CRtObserver
 {
     public:
-	    virtual void ontReceivingRt(const std::vector<Eigen::Matrix4d> &relative_transformations) = 0;
+	    virtual void ontReceivingRt(const std::vector<Eigen::Matrix4f> &relative_transformations) = 0;
 };
